Creating a maze array
Pair programming with Steve from codebar.io
I've started trying to make the maze into a 2D array so that PacWoman
and the food are all on the same grid. This is a test using a simple int
grid of 1s and 0s to denot e wall/not wall. Now I need to scale this up
and fit it to the map, and make it solid blocks of colour etc.